+++ pyriemann/utils/ajd.py
@@ -1,7 +1,8 @@
 """Aproximate joint diagonalization algorithms."""
 
-import numpy as np
 import warnings
+
+import numpy as np
 
 from .utils import check_weights, check_function
 
@@ -116,7 +117,7 @@ def ajd_pham(X, *, init=None, eps=1e-6, n_iter_max=20, sample_weight=None):
     Parameters
     ----------
     X : ndarray, shape (n_matrices, n, n)
-        Set of SPD matrices to diagonalize.
+        Set of SPD/HPD matrices to diagonalize.
     init : None | ndarray, shape (n, n), default=None
         Initialization for the diagonalizer.
     eps : float, default=1e-6
@@ -132,7 +133,7 @@ def ajd_pham(X, *, init=None, eps=1e-6, n_iter_max=20, sample_weight=None):
     V : ndarray, shape (n, n)
         The diagonalizer, an invertible matrix.
     D : ndarray, shape (n_matrices, n, n)
-        Set of quasi diagonal matrices, D = V X V^T.
+        Set of quasi diagonal matrices, D = V X V^H.
 
     Notes
     -----
@@ -166,6 +167,7 @@ def ajd_pham(X, *, init=None, eps=1e-6, n_iter_max=20, sample_weight=None):
         V = np.eye(n)
     else:
         V = _check_init_diag(init, n)
+    V = V.astype(X.dtype)
     epsilon = n * (n - 1) * eps
 
     for _ in range(n_iter_max):
@@ -187,7 +189,9 @@ def ajd_pham(X, *, init=None, eps=1e-6, n_iter_max=20, sample_weight=None):
 
                 tmp = np.sqrt(omega21 / omega12)
                 tmp1 = (tmp * g12 + g21) / (omega + 1)
-                tmp2 = (tmp * g12 - g21) / max(omega - 1, 1e-9)
+                if np.isrealobj(X):
+                    omega = max(omega - 1, 1e-9)
+                tmp2 = (tmp * g12 - g21) / omega
 
                 h12 = tmp1 + tmp2
                 h21 = np.conj((tmp1 - tmp2) / tmp)
@@ -195,15 +199,18 @@ def ajd_pham(X, *, init=None, eps=1e-6, n_iter_max=20, sample_weight=None):
                 decr += n_matrices * (g12 * np.conj(h12) + g21 * h21) / 2.0
 
                 tmp = 1 + 0.5j * np.imag(h12 * h21)
-                tmp = np.real(tmp + np.sqrt(tmp ** 2 - h12 * h21))
-                tau = np.array([[1, -h12 / tmp], [-h21 / tmp, 1]])
+                tmp = tmp + np.sqrt(tmp ** 2 - h12 * h21)
+                if np.isrealobj(X):
+                    tmp = np.real(tmp)
+                tau = np.array([[1, np.conj(-h12 / tmp)],
+                                [np.conj(-h21 / tmp), 1]])
 
-                A[[ii, jj], :] = tau @ A[[ii, jj], :]
+                A[[ii, jj], :] = tau.conj() @ A[[ii, jj], :]
                 tmp = np.c_[A[:, Ii], A[:, Ij]]
-                tmp = np.reshape(tmp, (n * n_matrices, 2), order='F')
+                tmp = np.reshape(tmp, (n * n_matrices, 2), order="F")
                 tmp = tmp @ tau.T
 
-                tmp = np.reshape(tmp, (n, n_matrices * 2), order='F')
+                tmp = np.reshape(tmp, (n, n_matrices * 2), order="F")
                 A[:, Ii] = tmp[:, :n_matrices]
                 A[:, Ij] = tmp[:, n_matrices:]
                 V[[ii, jj], :] = tau @ V[[ii, jj], :]
@@ -212,8 +219,8 @@ def ajd_pham(X, *, init=None, eps=1e-6, n_iter_max=20, sample_weight=None):
     else:
         warnings.warn("Convergence not reached")
 
-    D = np.reshape(A, (n, -1, n)).transpose(1, 0, 2)
-    return V, D
+    D = np.reshape(A, (n, -1, n)).transpose(1, 0, 2).conj()
+    return V.astype(X.dtype), D.astype(X.dtype)
 
 
 def uwedge(X, *, init=None, eps=1e-7, n_iter_max=100):